+++ src/main/java/com/wildcodeschool/skillhub/service/UserService.java
@@ -3,36 +3,18 @@ package com.wildcodeschool.skillhub.service;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import com.wildcodeschool.skillhub.model.User;
-import com.wildcodeschool.skillhub.repository.UserRepository;
 
-@Service
-public class UserService {
+public interface UserService {
 
-	@Autowired
-	private UserRepository userRepository;
+	Optional<User> findById(Long id);
 
-	public Optional<User> findById(Long id) {
-		return userRepository.findById(id);
-	}
+	List<User> findByuserSkills_SkillId(Long id);
 
-	public List<User> findByuserSkills_SkillId(Long id) {
-		return userRepository.findByuserSkills_SkillId(id);
-	}
+	List<User> findAll();
 
-	public List<User> findAll() {
-		return userRepository.findAll();
-	}
+	void deleteById(Long id);
 
-	public void deleteById(Long id) {
-		userRepository.deleteById(id);
-	}
-
-	public void save(User user) {
-		userRepository.save(user);
-	}
+	void save(User user);
 
 }
