+++ src/main/java/com/wildcodeschool/skillhub/service/UserServiceImpl.java
@@ -1,0 +1,43 @@
+package com.wildcodeschool.skillhub.service;
+
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import com.wildcodeschool.skillhub.model.User;
+import com.wildcodeschool.skillhub.repository.UserRepository;
+
+@Service
+public class UserServiceImpl implements UserService {
+
+	@Autowired
+	private UserRepository userRepository;
+
+	@Override
+	public Optional<User> findById(Long id) {
+		return userRepository.findById(id);
+	}
+
+	@Override
+	public List<User> findByuserSkills_SkillId(Long id) {
+		return userRepository.findByuserSkills_SkillId(id);
+	}
+
+	@Override
+	public List<User> findAll() {
+		return userRepository.findAll();
+	}
+
+	@Override
+	public void deleteById(Long id) {
+		userRepository.deleteById(id);
+	}
+
+	@Override
+	public void save(User user) {
+		userRepository.save(user);
+	}
+
+}
